[EDD] WooTube 3주차 기능구현 (#33)
* [feat] get current user api (#151)

* 현재 로그인한 사용자 정보를 가져오는 api입니다.

* [feat] add view count to video and change to retrieve with it

* [feat] impolement update view count

* [fix] fix issue with adding xss prevention to video publish request (#150)

비디오 게시 요청의 인자들이 XSS escaping을 거치도록 수정합니다.

* [feat] update video lsit and detali view to show view count

* [feat] update db schema for production to add view count column

* [fix] fix issue (#158)

- 중복 작성된 라인을 지웠습니다.

* [feat] add navi subcribe area

* [refactor] rename method

* [refactor] rename method common.js

* [refactor] change return type

* [feat] 사용자 정보 수정 및 탈퇴 API 연동 (#162)

* 웹페이지에서 사용자 정보 수정과 탈퇴가 동작하도록 합니다.
* app.css 파일을 교체합니다.

* [refactor] add console.log

* [feat] 내 채널 페이지에서의 비디오 조회 (#163)

* 채널 페이지에서 해당 사용자가 올린 비디오를 조회할 수 있도록 합니다.

* [refactor] add fetch - catch

* [refactor] change error response

* [refactor] change response

* [refactor] change err response

* [refactor] change err response message

* [refactor] update test to be based callback

* [fix] fix issue with request saving in order

* [refactor] change err response message

* [fix] resolve merge conflict

* [refactor] change comment

* [style] adjust method location

* [feat] apply peer feedback

* [feat] remove unnecessary line

* [refactor] rename service method

* [refactor] add column annotation to video entity

* [refactor] change url address

* [feat] add logback setting for file logging in production environment (#173)

- 프로덕션 환경에서 파일에 로그를 기록하는 설정을 추가합니다.

* [feat] CommentCRUD (#174)

* 코멘트의 생성 , 조회, 수정, 삭제 기능 구현
* 일부 테스트 코드 리팩토링

* [feat]add  connection logout Api (#176)

- 로그아웃 API 연동
- 헤더의 로그인 확인 기능 ( 로그인 한 상황 : 상단 헤더에 로그아웃, 로그인 하지 않은 상황 : 상단 헤더에 로그인 표시 )

* [fix] issue 168 (#177)

* 중국말 한국어로 패치, 코드 포매팅

* [feat] add front comment

- 댓글 작성/수정/삭제 마크업 및 ajax 통신 기능을 추가했습니다.

* [feat] add subscription (#180)

*  구독 API 구독 추가했습니다

* fix #182 (#185)

* 서버 응답의 날짜 포맷을 변경합니다.
* 이에 따른 프론트 페이지의 날짜 표기와 관련된 코드를 수정합니다.

* [fix] front index no header  (#186) (#188)

* 헤더와 사이드 바가 나오지 않던 문제를 해결했습니다.

* [refactor] change front comment delete btn (#193)

* 댓글 삭제, 수정 버튼 클릭을 더 수월하게 바꿨습니다.

* [refactor] add subscription schema (#194)

subscription schema를 추가했습니다.

* [feat] 나의 채널 페이지에서 나의 동영상들 조회 API (#196)

- 나의 채널 페이지에서 나의 동영상 조회하도록 API 수정

* [feat] add requestMyChannelVideos api in common.js,

* [refactor] change baseUrl

* [chore] delete conflict trace

* [refactor] delete duplicate common.js method

* [refactor] delete unnecessary blank

* [feat] add findbyCrator

* [feat] add front user create video

* [refactor] remove empty line

* [refactor] remove empty line

* fix #197 (#198)

* [fix] change source of bootstrap js/css

* [fix] change jquery cdn to jsdelivr

* [refactor] remove loading script integrity attribute

* [refactor] remove integrity attribute

* [fix] video user delete bug (#199)

- 탈퇴한 유저가 작성한 비디오를 가져올 때 발생하는 오류를 해결했습니다.

* [fix] Issue/201 (#202)

* 같은 이메일로 가입시 익셉션
* 이메일, 이름, 비밀번호 null시 익셉션

* [fix] fix to check whether if password matches its confirmation (#204)

* 회원가입시 비밀번호 확인도 함께 확인하도록 변경합니다.

* [fix] password check fail tst (#210)

*컴파일 에러 이슈 해결

* [fix] password confirm password (#212)

* 비밀번호 확인 validation 체크시 프론트에서 요청을 해주지 않았던 이슈 해결

* [refactor] change front video detail page style (#211)

- 비디오 상세 페이지 본문과 댓글 영역 스타일을 조정합니다.

* [feat] change to show videos in view count order (#213)

- 동영상 조회수 순으로 노출하는 기능 추가했습니다.

* [feat] feature/subscription button (#214)

* 프론트엔드 기능 구현완료

* [refactor] change front user leave btn issue (#208) (#215)

- 프론트에서 회원 탈퇴 버튼 클릭시 무조건 로그인한 사용자가 탈퇴되는 현상을 해결했습니다.

* [refactor] add email duplicate error (#217)

탈퇴한 이메일로 회원가입 시도시 DB 에러 캐치했습니다.

* [fix] resolve issue 228 (#230)

- comment, user XSS 추가했습니다.

* [fix] fix subscription bug (#232)

- 구독관련 함수를 common.js에 추가했습니다.

* [fix] fix duplicatEmailSignUpException (#233)

- 이메일 중복 메세지 내용을 변경했습니다.

* [fix] video authenticate (#220) (#234)

- 비디오 수정/삭제시 현재 로그인 사용자를 확인하는 로직을 추가했습니다.

* [feature] fix cors issue (#235)

- CORS 관련 필터를 추가하여 비로그인 상태에서 로그인이 필요한 요청 시 발생하는 401 응답을 기존 응답과 동일하게 처리할 수 있도록 수정합니다.
- 프론트엔드의 일부 불필요한 코드를 정리합니다.

* [fix] update video comment count #221 (#236)

- 비디오 상세 페이지에서 댓글 작성과 삭제마다 댓글 개수를 갱신하도록 합니다.

* [fix] change front video detail comment area markup  #222 (#237)

- 댓글의 수정/삭제 버튼이 작성자에게만 보이도록 수정합니다.

* [fix] #226 회원가입 비디오 게시 회원 수정 관련 size 검증 (#238)

* 회원가입 수정 게시 관련 size 검증

* [fix] change email validation #229 (#239)

email validation 추가.
기존 : ~~ @ ~~~
변경: ~~ @ ~~~ . ~~~ 가 되도록.
변경에 따른 테스트 코드 추가

* [fix] fix message for name (#231)

- 닉네임이 영어, 한글이 가능하다고 변경

* [fix] fix #223 (#242)

동영상 상세 페이지의 동영상 수정/삭제 버튼이 작성자에게만 보이도록 변경합니다.

* [refactor] change UserSaveRequestDto Name message (#245)

회원가입시 이름 에러 메세지를 변경했습니다.

* [fix] change front channel page (#247)

- 비로그인상태거나 자신 이외의 사용자가 채널 페이지에 접근하는 경우 회원정보 수정과 회원 탈퇴 버튼을 노출하지 않도록 변경합니다.

* [fix] add xss to update dto  (#249)

- domain을 수정할 때도 xss 방어 로직하도록 수정했습니다.

* [refactor] change front subscription area(#244) (#250)

- 구독 신청/취소 시 사이드 네비의 구독목록 리스트에 바로 반영되게 변경했습니다.
- 사이드 네비의 라이브러리 부분의 이름을 로그인 했으면 본인 이름으로 노출되게 변경했습니다

* [fix] change mypage video (#253)

- 채널페이지에서 비디오 리스트를 가져올 수 있도록 변경했습니다.

* [refactor] feedback (#248)

준 피드백 반영

* [feedback]feedback/252 (#254)

* [refactor] change front main page video count (#258)

+++ frontend/js/video/detail.js
@@ -1,31 +1,57 @@
 const detailView = function() {
-    params = wootubeCtx.util.getUrlParams();
-
-    const videoId = params.id;
+    const videoId = wootubeCtx.util.getUrlParams('id');
     api.requestVideo(videoId)
-    .then(response => response.json())
-    .then(data => detailVideo(data))
+        .then(response => {
+            if (response.status !== 200) {
+                response.json()
+                .then(err => {
+                    alert('해당 동영상이 없습니다.');
+                    window.location.href = "/"
+                })
+            } else {
+                return response.json()
+            }
+        })
+        .then(data => detailVideo(data))
+        .then(data => {
+            const btnEdit = document.querySelector('#btn-edit')
+            const btnDelete = document.querySelector('#btn-delete')
+
+            if (btnEdit) {
+                btnEdit.addEventListener('click', () => {
+                    location.href = `/video-edit.html?id=${wootubeCtx.util.getUrlParams('id')}`
+                })
+            }
+
+            if (btnDelete) {
+                btnDelete.addEventListener('click', () => {
+                    api.deleteVideo(wootubeCtx.util.getUrlParams('id'))
+                        .then(response => deleteVideo(response))
+                })
+            }
+        })
 }
 
 const detailVideo = function(data) {
-    if (data.result) {
-        alert('해당 동영상이 없습니다.');
-        return false;
-        // location.href = '/index.html';
-    }
-    addVideoDetailTemplate(data)
+    return new Promise((resolve) => {
+        api.retrieveLoginInfo()
+            .then(res => {
+                if (res.status === 200) {
+                    return res.json()
+                }
+            })
+            .then(json => {
+                if (json) {
+                    addVideoDetailTemplate(data, data.creator.id === json.id)
+                    return resolve()
+                }
+                addVideoDetailTemplate(data, false)
+                return resolve()
+            })
+    })
 }
 
-detailView();
-
-document.querySelector('#btn-edit').addEventListener('click', function(event) {
-    location.href = `/video-edit.html?id=${wootubeCtx.util.getUrlParams().id}`
-})
-
-document.querySelector('#btn-delete').addEventListener('click', function(event) {
-    api.deleteVideo(wootubeCtx.util.getUrlParams().id)
-        .then(response => deleteVideo(response))
-});
+detailView()
 
 const deleteVideo = function(data) {
     if (data.status === 404) {
